Introduce TerminalQueryExpression.
